moved workouts to correct directory
